requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
schedule>=1.2.0
